opencl: do not use constant memory in kernel parameters (could cause problems without giving advantages)
in addition:

* fix two hidden (but not virulent) bugs in sharpen and highpass

* disallow preview pipe to grab opencl device (might change again in
future if people typically should have more than one device)

* give a clearer final message about opencl availability when DT is
called with '-d opencl'

* do not try to load opencl kernel in module 'levels'

+++ src/iop/highpass.c
@@ -168,7 +168,7 @@ process_cl (struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, cl_mem 
   unsigned long localmemsize = 0;    // the maximum amount of local memory we can use
   
   // make sure blocksize is not too large
-  size_t blocksize = BLOCKSIZE;
+  int blocksize = BLOCKSIZE;
   if(dt_opencl_get_work_group_limits(devid, maxsizes, &workgroupsize, &localmemsize) == CL_SUCCESS)
   {
     // reduce blocksize step by step until it fits to limits
